+++ ICSSoft.STORMNET.UserDataTypes/WebFile.cs
@@ -12,7 +12,7 @@
     /// </summary>
     [StoreInstancesInType(typeof(Business.SQLDataService), typeof(string))]
     [Serializable]
-    public class WebFile
+    public class WebFile : IComparableType
     {
         /// <summary>
         /// Имя
@@ -59,6 +59,20 @@
                 stream.Seek(0, SeekOrigin.Begin);
                 return (WebFile)formatter.Deserialize(stream);
             }
+        }
+
+        /// <inheritdoc />
+        public int Compare(object x)
+        {
+            if (x != null && x is WebFile file)
+            {
+                if (file.Url == Url)
+                {
+                    return 0;
+                }
+            }
+
+            return -1;
         }
 
         public override string ToString()